Created a utility class called OpenUrl and fixed DRY problem.
