+++ src/main/java/ca/yorku/cse2311/tab2pdf/ui/listener/SampleInput_1Listener.java
@@ -1,6 +1,7 @@
 package ca.yorku.cse2311.tab2pdf.ui.listener;
 
 import ca.yorku.cse2311.tab2pdf.ui.MainJFrame;
+import ca.yorku.cse2311.tab2pdf.util.OpenUrl;
 
 import java.awt.Desktop;
 import java.awt.event.ActionEvent;
@@ -34,35 +35,10 @@ import java.net.URISyntaxException;
 
     @Override
     public void actionPerformed(ActionEvent e) {
-    	open(createURI(INPUT_FILE_1));
+    	new OpenUrl(INPUT_FILE_1);
     	
     }
-    /**
-	 * Opens the URI in the URI object using the desktop's default browser
-	 * @param URI uri
-	 */
-	public static void open(URI uri){
-    	if (Desktop.isDesktopSupported()){
-    		try{
-    			Desktop.getDesktop().browse(uri);
-    		}catch (IOException e){
-    			e.printStackTrace();
-    		}
-    	}else{
-    		//TODO
-    	}
-    }
+    
 	
-	public URI createURI(String link){
-		URI uri;
-		try{
-			uri = new URI(link);
-			
-		}catch (URISyntaxException m){
-			uri= null;
-			m.printStackTrace();
-		}
-		return uri;
-	}
     
 }
